添加额外破坏效果配置

+++ src/main/java/com/atsuishio/superbwarfare/entity/Tm62Entity.java
@@ -222,12 +222,13 @@ public class Tm62Entity extends Entity implements GeoEntity, OwnableEntity {
             }
 
             if (trigger) {
-                triggerExplode();
-                if (this.level() instanceof ServerLevel) {
+                this.triggerExplode();
+
+                if (this.level() instanceof ServerLevel && ExplosionConfig.EXPLOSION_DESTROY.get() && ExplosionConfig.EXTRA_EXPLOSION_EFFECT.get()) {
                     AABB aabb = new AABB(position(), position()).inflate(2);
                     BlockPos.betweenClosedStream(aabb).forEach((blockPos) -> {
                         float hard = this.level().getBlockState(blockPos).getBlock().defaultDestroyTime();
-                        if (ExplosionConfig.EXPLOSION_DESTROY.get() && hard != -1) {
+                        if (hard != -1) {
                             this.level().destroyBlock(blockPos, true);
                         }
                     });
